requirements.txt
aniso8601==7.0.0
asgiref==3.2.5
async==0.6.2
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
constantly==15.1.0
coverage==5.0.4
cryptography==2.8
daphne==2.4.1
Django==3.0.4
django-graphql-playground==1.0
gql-query-builder==0.1.5
graphene==2.1.8
graphene-django==2.9.0
graphene-subscriptions==1.0.2
graphql-core==2.3.1
graphql-relay==2.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==17.5.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
promise==2.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.8.0
pytest-pythonpath==0.7.3
pytz==2019.3
Rx==1.6.1
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
Twisted==19.10.0
txaio==20.1.1
wcwidth==0.1.8
zipp==3.1.0
zope.interface==4.7.2
docker
docker-compose
django-material-admin
